fsmc_nand.c: Support for erase block size of 256 KB.
The current patch is required to support EVALSPEAR1340CPU
Revision 2 where a new (ONFI compliant) MT29F8G08ABACA
NAND flash from Micron is present.

The geometry has a erase block of 256 KB, so we need to define
in the driver a proper partitioning scheme that takes this into
account.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@st.com>
Signed-off-by: Armando Visconti <armando.visconti@st.com>