nand/fsmc: Newly erased page read algorithm modified
An erased page could not be read properly with BCH8 algorithm because the ecc
for an erased page is not ff..ff. To resolve this problem, the page is compared
against ff..ff and correction is not performed.

A drawback of this algorithm is that when there are bit flips in an erased page,
the correction is tried which results in a read error. This means an erased page
with bit flips can not be read properly. This problem is solved by checking the
number of bits (in 512 byte data + 13 byte ecc) which are 0. If these number of
bits are less than 8, the page is considered erased and correction algorithm is
not tried on that page

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>