SPEAr: Update dma_data for users of dw_dmac
With the addition of DMA_SLAVE_CONFIG in dw_dmac driver, the way to use it is
changed. Now, some information is passed fixed for user drivers and some is
platform dependent. With this few fields of struct dw_dma_slave are now useless
and they must not be used. This patch has changes related to this for SPEAr SoC.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>