phylib: Change MDIO bus dev_pm_ops Callback
The Synopsis MAC requires all clocks to be on at the time of Software reset.
The thaw call back for the MDIO bus was being executed post the stmmac driver
thaw call back, which was trying to reset the Synopsis MAC IP. This returned
error from the stmmac_open routine.
The use of thaw_noirq and freeze_noirq ensured the call backs for MDIO freeze
and thaw are schedules post stmmac driver hibernation call backs. This ensures
the PHY is shutdown after GMAC and is resumed before GMAC, and hence forth the
25MHz clock from PHY is present at time of MAC reset.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>