media: soc-camera: Implement CROPCAP, G_CROP features and fix S_CROP implementation
This patch adds the following new features (which were missing earlier in
CAMIF driver):
1. CROPCAP
2. G_CROP

In addition this patch also fixes the S_CROP implementation in CAMIF driver.

As CAMIF by itself supports CROPPING, we first request CROP features
from the underlying subdev. If the subdev does not support CROPPING
or cannot CROP to a rectangle with sufficient precision, we invoke
the CROPPING feature of CAMIF.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>