SPEAr13xx PM: Shuffle the Invalidation code post DDR is up.
This patch moves all the invalidation related code post the initialization of
DDR. This assures that the system memory is out of self refresh, and then only
Cache related aspects are performed. This doest not have any significance
though as primarily only invalidation was being done, but the code looks much
clean now.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>