media: soc-camera: Fix current/active frame handling in CAMIF and locking mechanism
As the videobuf-core uses the 'lock' provided by CAMIF in 'sg_init' routine
to protect buffer insertion and deletion from the video-buf queue, the CAMIF
driver should also use the same 'lock' instance to protect the videobuf
insertion/deletion in its own routines.

Also the current buffer (which was/is the last ACTIVE buffer) notion needs
to be more simpler in the driver.

As per the DWC dma design, the dma 'do_start' functionality is performed at
the time of 'tx_submit' call and hence calls to 'issue_pending' are not
required.

This patch adds these three features.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>