media: soc-camera: Add SW workarounds for CAMIF hang during DMA-transfer
The CAMIF module in SPEAr1340 hangs while performing DMA bursts.
As per the HW designer this can occur when the CAMIF module has raised
a DMA BURST/SINGLE request but in response it received on DMA CLR signal.

In such a case, the only possible remedy is to perform SOFT RESET of
CAMIF and bring all the SW data structures in their correct state.

Secondly, there also be a problem caused by the ASYNC relationship
between when the frames are queued by an user-space application (via
QBUF ioctl) and when DMA completion callback corresponding to this frame
is called. This can lead to the classic producer-consumer problem in which
the application is not able to queue buffers fast enough.

This patch implements SW workarounds for both the issues discussed above.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>