SPEAr1340/LCAD: Add support for SPEAr1340_LCAD board
LCAD stands for "Low Cost Access Device". This patch adds support for
1340 based LCAD board.

Note:
pl061 gpio drivers programs all gpio pins in input mode by default. We
don't want GPIO0_3 (on this board) to be programmed to input mode as
this is required to supply power to board. This patch fixes GPIO0_3 to
stay in output mode val 0. Others are left in input mode.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>