USB/ehci-spear: Fix for EHCI resume.
The fix checks for the configured flag if already set, and restores the Host
configurations. This is specifically the case when the USB HC power domain
has not been made off.
In case the USB power domain is off, the configured flag read from the HW gets
reset and the EHCI is reset to cope with power off.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>