media: soc-camera: Add 'suspend' and 'resume' handlers for VS6725 sensor
The VS6725 sensor should support the 'suspend' and 'resume' handlers
provided by soc_camera_ops (Note that these are not dev_pm_ops specific
handlers).

During suspend the Chip Enable(CE) of the VS6725 sensor should be driven
LO to save power and after resume the same should be driven HI.

Note that as we have to program patches and default registers after
resume, we should keep a notion of the image size and format we
were working with before we went in suspended state. The same image
size and format should be programmed again before we set the sensor
in RUN mode.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>