media: soc-camera: Seperate QBUF producer/consumer mismatch issue from CAMIF HW bug
While processing videobufs provided by the user-application the CAMIF can
have the classic producer/consumer throttle mismatch issue where the application
is not able to queue the videobufs as fast as the CAMIF driver can process them.

In such cases it does not make sense to RESET the entire CAMIF module as we
do in case the DMA gets hung as a result of CAMIF HW issue. Instead we should
just reenable the Frame-End interrupts and wait for a new videobuffer to be
queued.

This patch adds this fix.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>