sound/codec/sta529: Change interface data length to 32 bit.
This change is specific to SPEAr320s only.
For audio to play/record, the reference clock required is 12.284Mhz and
with this clock we can generate sclk as 3.071Mhz as bitclock (sclk =
refclock/4 = 12.284/4 = 3.071Mhz, as there is a fix divider i.e 4). The
above generated bit clock is not for 16 bit sampled song. With this we
have to play/record 32 bit format song.

To overcome this limitation we need to fix (irrespective of data
format) STA529 register configuration i.e. serial-to-parallel
interface data length and parallel to-serial interface data length as 32 bit.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>