SPEAr13xx PM: Add Hibernation ops plugin
The poweroff() callback of individual drivers are called in the last phase of
hibernation, after the image has been created and the system is going for
"power off".
The ->poweroff() callbacks are only executed if hibernation_mode is equal to
HIBERNATION_PLATFORM, which is not the default case. This requires the
platform specific plug-in that should be registered into the kernel suspend to
disk framework.
This patch adds in the support for SPEAr platform specific hibernation plug-ins
which calls hibernation_set_ops() and sets the operations appropriately.
These operations have been kept empty as they don't need to do anything, but
have been defined so as to ensure the prerequisites for ->poweroff callbacks
have been met.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>