media: soc-camera: Fix 'select-timeout' issue when putting CAMIF in Hibernation
The CAMIF driver uses the legacy pm operations (suspend/resume) to support
Hibernation as well.

In case of invoking the suspend-to-disk power save mechanism,
the application accessing the CAMIF driver was throwing a 'select-timeout'
error on resuming from the saved image.

This was primarily because of the CAMERA module being in RESET state or
the frame-end being disabled even for the 1st frame from the sensor
after RESUME.

This patch fixes this issue and now CAMIF is working fine for Standby,
Suspend-to-RAM and Suspend-to-Disk power-save mechanisms.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>