SPEAr/VDEC: Add Memalloc driver support.
This patch adds support for Memalloc driver used by SPEAr video decoder/encoder
to reserve contiguos physical memory space. Presence of this driver is
mandatory to enable decoder/encoder to work in a Linux system.  A default 64M
memory table is selected, starting address of memory area is 256M to be able to
work together with default Mali/UMP configuration that reserve the area between
128M and 256M.

Signed-off-by: Salvatore De Dominicis <salvatore.dedominicis@st.com>