media: soc-camera: Disable SINGLE DMA requests from CAMIF
As per the discussion with the HW designers the SINGLE DMA request feature
cannot be used with the V4L2 framework's way of programming the LLI's for
each frame.

So, this patch disables the SINGLE DMA requests from CAMIF and if the
resolutions selected from the user-space (via S_FMT/S_CROP commands) leads
to a case where we can have a SINGLE requests at the end of the transfer then
we simply return with a WARNing message.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>
Reviewed-by: Bhavna Yadav <bhavna.yadav@st.com>