media: soc-camera: Fix CAMIF power-management issue
This patch fixes the Issue #156041 logged in CODEX corresponding
to power-management features (S2R and standby) failing on CAMIF
driver when the tests were executed multiple times with the help
of a script.

The issue with the power-management implementation in CAMIF driver
was the missing case of handling the case when the application
has QUEUED a video-buffer and before the same could be DEQUEUED,
CAMIF was suspended.

In such a case the CUR_FRM is not NULL and as soon as CAMIF resumes,
we should handle the already present CUR_FRM. This is done by simply
enabling the FRAME_END interrupt if the CUR_FRM is not NULL. Otherwise
we need to wait for new video-buffer to be QUEUED before enabling the
FRAME_END interrupt.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>
Reviewed-by: Bhavna Yadav <bhavna.yadav@st.com>