input/lsm303dlh: enable_irq irrespective of regulator absence
The driver enabled irqs only in case the power regulator is present on the
device. This makes the driver dormant for boards/machines where regulator isn't
present. Since a regulator is an optional component, this makes sense to
separate regulator access and enabling irqs. Hence, this commit calls enable_irq
irresepective of presence or absence of the regulator

Signed-off-by: Amit Virdi <amit.virdi@st.com>
Reviewed-by: Viresh Kumar <viresh.kumar@st.com>