SPEAr1340: Optimize cpufreq transition latency.
Currently cpufreq transition latency value does not really reflect the
actual SPEAr1340 PLL transition delay. This patch has the actual latency
value, based on the PLL specification plus an arbitrary adding of 100 us
suggested by chip designers in order to be sure that the transition is
terminated. With this change, performance of conservative and ondemand
governor are improved.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@st.com>
Reviewed-by: Vipin Kumar <vipin.kumar@st.com>
Reviewed-by: Viresh Kumar <viresh.kumar@st.com>