stmmac: Define the option for DMA burst mode.
The stmmac core DMA could be programmed for Fixed burst or no
fixed burst option.
- The fixed burst configuration requires programming of the permissible
fixed burst length in newer versions (AXI supported) of stmmmac. This
AXI_BUS_MODE) register is reserved for earlier versions of stmmac and
writing to them has no impact.
- The no fixed burst option requires to program the maximum allowable burst
length for any stmmac core with AXI support.

This patch provides the configuration support for the fixed burst mode and
burst length used in the driver. The platform code need to set the fixed_burst
bit to enable the fixed burst mode.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>