stmmac: Add support for CPU freq notifiers
This patch adds in the support for CPU freq notifiers. In case the system freq
is changed using the CPU freq governors, the CSR input clock is also affected.

To maintain the CSR mdio clock within the limits of 1-2.5 Mhz the input clock
has to be properly scaled using the appropriate scaling factors.
This patch looks into the input clock frequency changes and selects the apt
scaling factor once a notifier alert is received.
Additionally the stmmac clock has been defined to support cpu freq notifiers
The MDC clock could be set higher than the IEEE 802.3 specified frequency limit
of 2.5 MHz to 12.5 MHz for the interfacing chips that support higher MDC clock.
This would require a further enhancement in the code.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>