stmmac: Adjust frame length for Type-1 Rx COE
The stmmac core supporting Type-1 Rx checksum offload engine (COE) append the
HW checksum at the end of payload in case the Rx checksum engine was used to
offload the HW checksum.
These cores did not provide the HW register interface to read the device
capabilities, and hence the plat code provides the core checksum capabilities
that help to identify type-1 cores, and adjust the frame length.

This patch addresses the adjustment of Rx frame length for the Type-1 RX COE.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>