thermal: spear_thermal: add Device Tree probing capability
SPEAr platforms now support DT and so must convert all drivers to support DT.
This patch adds DT probing support for SPEAr thermal sensor driver and updates
its documentation too.

Also, as SPEAr is the only user of this driver and is only available with DT,
make this an only DT driver. So, platform_data is completely removed and passed
via DT now.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>