SPEAR13xx: Add arch specific hook for gic set wake
This patch adds the platform specific hook for gic_set_wake.
This helps to address the following issue:
If enable_irq_wake fails, and then without tracking the return status
the disable_irq_wake is called, the unbalanced irq warning is obtained.
The status check for enable_irq_wake has been discouraged in the
drivers. Follow the link for discussions:
http://www.spinics.net/lists/linux-input/msg19475.html

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>