SPEAr6xx: stmmac: Update the platform initializations for stmmac core.
This patch updates the stmmac gmac plat code configurations based on changes
done in the stmmac driver and the HW capabilities of the SPEAr36xx stmmac core.
Since there are variations in GMAC core and PHY devices (PHY address etc), the
board specific initialization incorporates the platform specific init for the
core.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>
Reviewed-by: Viresh Kumar <viresh.kumar@st.com>