SPEAr13xx: stmmac: Update the platform initializations for stmmac core
The stmmac core configuration varies from board to board within SPEAr13xx
family. This patch updates in the support for the board specific platform data
for the stmmac core configurations based on the stmmac driver requirements and
stmmac core HW capabilities.
Along with that, certain stmmac core require explicit clock setups for the PHY.
The patch also updates the callbacks related to these configurations based
on PHY clock con_ids provided in the clock framework for the PHY's.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>
Reviewed-by: Viresh Kumar <viresh.kumar@st.com>