stmmac: Change mtu size w/o getting n/w down
This patch adds in the support to change the MTU size on the run once
the n/w interface is up.
Earlier the n/w had to be brought down and then only the MTU size could
be changed, but with this patch there is no need to bring the n/w down.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>