spi:pl022: enable interface clk in pl022_remove()
For every device two separate clocks are required: Functional and Interface
clock. In our case these two clocks are controlled via a single bit, and so we
have only one struct  clk for it.

AMBA bus enables interface clock in amba_probe before calling pl022_probe() and
it expects that this clock is still enable after pl022_remove() is called.  But
in case of spi-pl022 we disable interface clock at the end of probe to save
power.  So we need to enable interface clock in pl022_remove().

Also, functional clock is enabled/disabled around spi transfers and so it will
be disabled when pl022_remove() is called. So no need to disable it there.

Signed-off-by: Vipul Kumar Samar <vipulkumar.samar@st.com>