SPEAr1310 PM: Add support for suspend to ram for 1310 Rev-c
The patch adds the support for 1310 Rev-C. It includes following changes
1. The low level code for handling sleep has been modified for 1310
rev-c. This has been specifically done due to large changes in the
register configurations and bit space.
2. Corresponding changes to ensure proper calls for low level handlers.

NOTE: The low level routines work fine and have been tested
functionally. Due to the absence of GPIO patches on the board, the
system validation is still incomplete.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>