SPEAr1310 Revc PM: Modify the code for new HW register alignments.
This patch takes care of the following:
1. The configuration register bits of 1310 Rev-c differs from all other
SPEAr13xx architectures. The patch aligns the code with the HW changes
done in registers.
2. Further the Power domain switch off sequence is corrected in the code.
The code is iterated during the suspend to ram process.

NOTE:
The USB as a wake up source has been disabled, as enabling the USB wake
up source does not allows the SPEAr to move into suspend.  Also, the
PCIE wakeup configuration has still not been used to be in line with
the wake up code across 13xx family.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>