stmmac: Fix for nfs hang on multiple reboot
It was observed that during multiple reboots nfs hangs. The status of
receive descriptors shows that all the descriptors were in control of
CPU, and none were assigned to DMA. Also the DMA status register
confirmed that the Rx buffer is unavailable.  This patch adds the fix
for the same by adding the memory barriers to ascertain that the all
instructions before enabling the Rx or Tx DMA are completed which
involves the proper setting of the ownership bit in DMA descriptors.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>