sound:asoc:designware: Move i2s register initialization from hw_params to i2s_start.
Previously i2s register initialization are performed in hw_params function. It
will create issue in case of suspend to ram and suspend to disk. In both the
situation, since the power domain are off, all the register becomes reset. In
case of hibernation and suspend to ram, soc core will not call hw_param
function and so registers will not get initialized. So moving register
initialization to i2s_start.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>