SPEAr 1310-C: Touchscreen: Use gpio pin for spi chipselect
The touchsceen based on 13xx base expansion board is spi-based.  We can
use either gpio pins as chipselect or the chipselect pins provided by
the SSP IP.

As the hardware changes on the expansion board are made for gpio pins,
make corresponding changes in the board file for ssp chipselect option.

Signed-off-by: Vinit Shenoy <vinit.shenoy@st.com>