ASoC:designware: handle SPEAr 320s for configuring CCR register differently
SPEAr320s has a fix divider (div 4) to generate sclk from ref clock. Due
to this the generated sclk clock (which must be ref clk/8) is double
than the desired clock. As a result we cannot play or record 16 bit
sample size data in SPEAr 320s.

As a workaround we take SPEAr320s as a special case and explicitly
program i2s controller for 32bit sample width in this case.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>