spear320s: audio: workaround: Change sound format reported to ALSA to 16 bit.
There has been a strange observation on SPEAr320s audio. When we play
32bit sample audio data and accordingly configure I2S controller, clock
and the codec we are not able to observe correct audio timings and the
play data is stretched.

The cause for this is still unknown and under debug. For now we export
support for 16 bit data to ALSA (which is wrong) and controller, clock
and codec remain configured for 32bit. Surprisingly, this makes audio
work properly.

This patch will make things work, till further observation is made on
this.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>