SPEAr: Implement sched_clock for spear
This adds the sched_clock() override for the weak in-kernel
jiffybased sched_clock(). Thus giving better time resolution.

Please note that the kernel starts using sched_clock() little early in
the boot sequence when clock source is not even initialized. Protect it
in such cases by checking if it is really initialized.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>