SPEAr1310: Add platform support for stmmac
This patch adds the support to pass the plaform specific data through
DEV_AUX_DATA.

As of now, the support has been provided through this mode as the DT
does not have a  support for the callbacks that are required for
configuration of clocks required by different ethernet interfaces which
are more SOC specific.

As the DT matures and the clock framework is associated more closely
with DT, this functionality would be moved to the apt place.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>