SPEAr320: Add timer support required for macb HW fix resolution
This patch adds in the support to provide hardware timers related
callbacks that would allow to configure and run the timers.

The usage is more specific to macb, which has a HW issue and requires a
software workaround to flush the transmissions fifo in case the Tx
interrupts are not received in stipulated time intervals.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>