macb: Add driver support for SPEAr320 machine
This patch adds in the driver support for macb. Though the standard
driver exists for macb, but due to some HW bug a software workaround has
been used to fix the issue.
This workaround deviates the code from normal, and hence forth a
separate commit for the macb driver has been made.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>