spi/pl022: Fix the dma map/unmap calls sequencing
The current driver implementation for the DMA tranfers first invalidates
the cache for Rx buffers and then cleans the cache for Tx buffers before
starting the DMA transfers.

The SPIDEV framework assigns a single buffer for both reception and
transmission of data.So the above sequence of invalidating the cache and
then cleaning it does not work well when the same buffer is used for Rx
& Tx.

This patch alters the sequence, i.e first flushes the cache and then
invalidate them to ensure data integrity.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>