pinctrl/spear: workaround to access plgpio base
Pinctrl and plgpio are implemented through two separate drivers, one
being part of standard pinctrl framework and the other in the ususual
gpiolib framework.

However pinctrl needs to know the base of the allocated gpio in
order to translate to the actual pin required to be used as
a gpio.

Presently there is no formal mechanism to do this. We employ here a
dirty workaround by taking the gpio_chip in a global variable to be
referenced by pinctrl implementation.

Pinctrl also relies on probe of gpio to be over to allocate the gpio
ranges which it can serve. Hence if it doesn't find gpio chip to be
initialized it simple requests for deferred probing.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>