i2c/designware: Do not use _interruptible_ variant call for wait
If there is a signal pending and wait_for_completion_interruptible_timeout
exited because of the -ERESTARTSYS error we are unable to send any more i2c
messages.

Further this would lead to unsafe exit of i2c xfer as some of the variants
expect software to send stop command along with the last data, to mark the end
of the transfer.

So, use wait_for_completion_timeout instead.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>