SPEAr1340: Configure cpu at 600MHz driven by pll1_clk
As per simulation result pll3 is not good for cpu_clk and fails on slow
part. Other solution is to fix vco1/ pll1 on 1.2GHz and feed cpu_clk
through pll1.

This patch is to move cpu_clk on pll1_clk which is fixed at 1.2GHz. Also
the ahb bus is now clock asynchronously through the synthesizer at 166
MHz. This ensures that the bus frequency remains at 166 Mhz even if the
CPU frequency is changed using the cpu-freq.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>