SPEAr13xx: create strongly ordered memory mapping for sram
This patch defines memory to be used for storing and executing the low
level code for power management as strongly ordered memory.

The current low level handlers have fixed certain sections in the memory
as code, stack and data segments; and this has been essentially kept as
static and requires careful handling.

Besides the advantage of having the section as non cacheable/bufferable,
the mapping allows us to handle the low level code as per our design.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>