SPEAr13xx PM: Add wakeup and sleep low level handlers.
This patch adds in the machine specific code to handle the suspend to
ram and standby. The code is reallocated to SRAM by the code in pm.c.

The code maps the SPEAr architecture specific Power Management into
Suspend to RAM and Standby implementations at low level and takes care
of putting the system into low power mode.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>