SPEAr13xx PM: Add support for Suspend to RAM and Standby modes.
This patch provides the machine specific plugins required for Suspend
to RAM and Standby mode on SPEAr13xx platform.
The patch copies the machine specific handler for moving the system
into sleep to SRAM.
The Standby & Suspend to RAM handling differs primarily on the use of PM
block in SPEAr13xx platform which disables all the power islands for the
Suspend to RAM handling and allows the system wake up through selective
sources only.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>