SPEAr3xx/6xx: Add suspend to ram and standby support.
This patch adds in the suspend to ram and standby support for SPEAr3xx
and SPEAr6xx machines. The low power states from machines perspective
is same for both the suspend and standby implementations, and ensures
system goes into the sleep mode with DDR in self-refresh.

The sleep specific handling is done through Always on RAM, and this code
the patch carries forwards this copy operation of code into SRAM.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>