SPEAr13xx: Hibernate: Add the platform support
This patch set adds in the following support:

1. Initialize pm_power_off: For now the implementation has been an
infinite loop, the user can himself use gpio to pull system out of
hibernate.  This could well be changed to switching off the power
domains.

2. The poweroff() callback of individual drivers are called in the last
phase of hibernation, after the image has been created and the system is
going for "power off".

The ->poweroff() callbacks are only executed if hibernation_mode is
equal to HIBERNATION_PLATFORM, which is not the default case. This
requires the platform specific plug-in that should be registered into
the kernel suspend to disk framework.

This patch ensures that the prerequisites for ->poweroff callbacks have
been met, and adds dummy platform call back handlers for hibernation.
Though there is no specific low power state particular to platform, this
helps to trigger the device low power states before the machine could be
made off.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>