video:db9000: Implement runtime pin control in driver
To save power and damage to the panel we can dynamically configure clcd pins as
gpio or to the ip. While going into suspend mode configure clcd pins to gpio and
get back these pins to ip in resume.

This patch it to runtime configure clcd pins to save panel damage.

Signed-off-by: Vipul Kumar Samar <vipulkumar.samar@st.com>