stmmac: disable mmc interrupt status read
MMC is an optional hardware block, and the MMC interrupts were
disabled in the code even for the HW suppporting MMC module.

Since the register being read is a interrupt status register, if the HW
supports MMC, and the counters have not been freezed at start in the MMC
control settings, this bit would get set on some event condition in MMC.

The Normal and abnormal interrupts in GMAC also cause the interrupt
handler to trigger, and since the status of the bit would reflect MMC
counter updates the status of the MMC interrupt would always be printed
on console even if MMC interrupts are disabled.

MMC is currently being used to provide net stats using ethtool, and the
MMC interrupts have not been catered, this patch disables the MMC
interrupt status checks.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>