STMAC: Update driver to support various synopsis cores.
This patch set modifies the driver to support checksum offload engine
(coe) on older synopsis cores. Further since the Rx COE is optional and
configurable, there have been driver updates for core with no COE
support. The changes done in the driver are:

1. For the Synopsis GMAC Cores, the Frame status with full checksum
offload engine enabled have couple of status checks which are not
backward compatible for Version 3.4 and below.

These status checks correspond to
	A. non IPv4 and IPv6 frames
	B. Ipv4 and Ipv6 frames with no IP HCE and payload check bypass
	due to unsupported pay load.

The code has been modified not to discard frames for the above two
conditions for older cores.

2. Certain return status of the error frames have been added.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>