phylib: Change MDIO bus dev_pm_ops Callback
The Synopsis MAC requires all clocks to be on at the time of Software
reset.  The thaw call back for the MDIO bus was being executed post the
stmmac driver thaw call back, which was trying to reset the Synopsis MAC
IP. This returned error from the stmmac_open routine.

The patch makes use of thaw_noirq and freeze_noirq and ensures the call
backs for MDIO freeze and thaw are scheduled post stmmac driver
hibernation call backs.

Now the PHY is shutdown after GMAC suspend and is resumed before GMAC
resume and hence forth the 25MHz clock from PHY is present at time of
MAC reset.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>