drivers/clk : Modify clk settings for ADC
This patch contain:

1. Add correct devID during registration of ADC gate clk and ADC clk lookup.

2. Add CLK_SET_RATE_PARENT flag while registering ADC clock.

CLK_SET_RATE_PARENT flag is needed by children of respective nodes that
depend on parents for the rate settings.

For example, ADC requires the clock to be set to some value at run time.
Since ADC as a device depends on top level framework clock driven by
Synthesizer it propagates the control to the synthesizer only when the
above defined flag is set.

Further more, the hierarchy defined in the clock FW is as follow for
SPEAr ADC device.
        AHB --> Aux Synth --> Aux Synth Gate Clock --> ADC clock.

So for the ADC clock to be set to a certain value, the control goes to
ADC synth gate clock.

This patch sets the CLK_SET_RATE_PARENT for the gate clock also so the
clock set call could propagate rate change up one level to the
synthesizer which could effectively set the desired rate.

Signed-off-by: Vijay Kumar Mishra <vijay.kumar@st.com>