SPEAr13xx: Correct the filter function for DMA
The filter function call back implementation of the DMA breaks because
there is no provision to associate the slave devices with the DMA
controllers.

The filter function checks the respective association before allocating
the channels, and this association was previously passed through the
platform data of the slave devices.

With DT coming up, there is no way the slave devices could be associated
with the DMA controllers like it had been done before.

This patch set makes use of device id's assigned to the DMA controllers
and then associates the slave devices with DMA controllers by matching
those device ids in the filter function callback.

The DMA controller device ID are also passed using platform data, and
are assigned to the DMA controller device structures maintained by dma
engine by the DMA driver.

Additionally this patch also provides the corresponding dma master id's
to the slave to map on to the dma host controller

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>