ARM: SPEAr13xx: Add ADC channel0 DMA support.
This is an incremental patch to pass DMA controller device ID for
DMA operation of ADC channel0.

ADC channel0 support DMA operation which needs DMA controller device
information.

Earlier, it was passed as dma_dev from the platform data.
Now, due to DT implementation and corresponding cleanup it should be
passed through dma_master_id to map on to the dma host controller.

Signed-off-by: Vijay Kumar Mishra <vijay.kumar@st.com>