stmmac: Freeze mmc counter
MMC is an optional hardware block, and the MMC interrupts were
disabled in the code even for the HW suppporting MMC module.

As per the GMAC IP Documentation, the gmac core interrupt would be
generated even if the MMC interrupts are masked as and when the mmc
related status is updated in the DMA status registers.

This patch freezes the mmc counters during the init tim, as a temporary
fix for the above mentioned problem. The solution would be catered post
discussion with the mainline.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>